backend/requirements.txt
Flask
flask-cors
numpy
Keras-Preprocessing
pillow

opencv-python-headless
Pywavelets
